Attempt to improve long RTT catchup time during stream moves.
